# standalone/portal-api/src/modules/auth/dto/update-user-profile.dto.ts
import { ApiPropertyOptional, OmitType } from '@nestjs/swagger';
import { ValidateIf } from 'class-validator';
import { UpdateUserDto } from 'src/modules/auth/dto/update-user.dto';

/**
 * update user profile transfer object
 */
export class UpdateUserProfileDto extends OmitType(UpdateUserDto, [
  'status',
  'roleId',
] as const) {
  @ApiPropertyOptional()
  @ValidateIf((object, value) => value)
  avatar: string;
}