# standalone/portal-api/src/modules/refresh-token/entities/refresh-token.entity.ts
import {
  BaseEntity,
  Column,
  Entity,
  Index,
  PrimaryGeneratedColumn,
} from 'typeorm';

@Entity({
  name: 'refresh_token',
})
export class RefreshTokenEntity extends BaseEntity {
  @PrimaryGeneratedColumn()
  id: number;

  @Column()
  userId: number;

  @Column()
  ip: string;

  @Column()
  userAgent: string;

  // @Index()
  @Column({
    nullable: true,
  })
  browser: string;

  @Column({
    nullable: true,
  })
  os: string;

  @Column()
  isRevoked: boolean;

  @Column()
  expires: Date;
}